Move zstd_path fixture to test_file_utils
